Added SavingsAccount and updated related tests
